backend/app/requirements.txt
python-decouple==3.4
Django==3.2.10
django-environ
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
Pillow==8.4.0
gunicorn==20.1.0
whitenoise==5.2.0
django-cors-headers==3.7.0
djangorestframework_simplejwt==4.7.1
django-storages==1.11.1
boto3==1.17.88
psycopg2
django-rest-auth==0.9.5
django-allauth==0.44.0
redis==3.5.3
djangorestframework-gis==0.17
django-filter==2.4.0
APScheduler==3.7.0
